Moved public methods above private methods in the Cohort class.
Added a let block to the cohort_spec.rb file so declare values
of variables that should be compared in tests.

Changed name of ABtest (and ABtest_spec) to ab_test (ab_test_spec).
