+++ data_science/cohort.rb
@@ -15,6 +15,11 @@ class Cohort
     @std_dev = compute_std_dev
   end
 
+  def compute_confidence_interval_95pct
+    interval = @std_dev * 1.96
+    [@conversion_rate - interval, @conversion_rate + interval]
+  end
+
   private
 
   def get_cohort_name(population_data)
@@ -25,8 +30,6 @@ class Cohort
     @name = current_cohort.to_sym
   end
 
-  private
-
   def get_num_conversions(population_data)
     @num_conversions = 0
     population_data.each do |entry|
@@ -35,18 +38,11 @@ class Cohort
     @num_conversions
   end
 
-  public
-
   def compute_conversion_rate
     @conversion_rate = @num_conversions.to_f / @sample_size
   end
 
   def compute_std_dev
     @std_dev = Math.sqrt(@conversion_rate * (1 - @conversion_rate) / @sample_size)
-  end
-
-  def compute_confidence_interval_95pct
-    interval = @std_dev * 1.96
-    [@conversion_rate - interval, @conversion_rate + interval]
   end
 end
