+++ data_science/spec/ab_test_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../ABtest'
+require_relative '../ab_test'
 
 require 'json'
